chore(test): refactor resource type handler test
